languages/golang: in development, use an Alpine-based image with bash.
Eventually we want to exclusively use ephemeral containers, but they're
not widely available yet. So meanwhile, we allow a divergence in the
base images between dev and prod, so that we have tools available in the
dev image, available for development. The testing image remains
pristine, which should help identify situations where the server may
inadvertidely implicitly depend on the environment.

Fixes #147.