console: reduce the amount of updates, by checking line-by-line changes.
This helps in scenarios where we don't have additional logging being
emitted, as with --log_actions or just generally build outputs, etc. In
those cases, because we're printing them above the action block, we
always need to re-print the action block so it's placed at the right
spot.

This may lead to considering whether console.Output should be inlined so
minimize some of these redraws.

For #146.