build/buildkit: if target name is known, push image to registry directly from buildkit.
This removes the need to download the image from buildkit, which for nodejs/web
images saves multiple seconds.

The downside of this change is that a buildkit build now blocks on a name being
allocated, and no longer runs in parallel to e.g. ecr.make-repository. But that
seems like a small cost to pay, to get back multiple seconds back.

Fixes #265.