workspace/source/protos: allow google/api imports, as well as namespacelabs.dev/foundation imports.
This was changed from the catch-all function as it was leaking any proto file
that was compiled into foundation, but with the problem that those used protos
would be versioned with the binary. So constrained the lookup on google/api
protos, and added ability to load foundation protos from the dependency path
(rather than the compiled in ones).