experimental: added support to write deno-based deferred invocations.
Previously, deferred invocations had to be containerized binaries, which add
considerable build cost, and due to the container preparation time lead to
longer runtimes.

This is a first exploration, and is not yet available for broader user.

However, our requirements for tools are simple: share-nothing, mostly
no-network, low latency invocation with short duration. Deno is a possibility
that fits the bill (WASM is another exploration path, and Deno may be an
enabler there as well). Deno is particularly attractive because very much like
`ns`, it's a single binary distribution. Which makes it trivial to build on,
unlike nodejs which is dependency heavy.