Make sure that cluster provider can be correctly deserialized from devhost.
After d888765 nscloud started to expect ProviderSpecific to include the cluster ID. However it only got filled if the cluster was actively created, not if it was deserialized from devhost. This caused the following error:

% ns prepare new-cluster
% ns dev server/resultsweb
Failed: internal error: cluster creation state is missing

This change makes sure ProviderSpecific is filled with cluster ID from devhost. The actual type of ProviderSpecific is being changed from CreateClusterResult since we cannot reproduce it from devhost reliably (and we don't really need it).