Implementing filesync for opaque dockerfiles. (#940)
Syntax:

mounts: "/app/src": {
	syncWorkspace: fromDir: "src"
}
Implementation:
If there is a "syncWorkspace" mount:

A layer with the "filesync-controller" is added to the built image.
ns dev uses a custom "hot reload" module that forwards filesync events to the dev observer (same way it works for node.js/web today).
The dev observer starts the filesync controller process inside the same container. When ns dev exists, it sends Ctrl-C to the controller process. As a failsafe, ns kills all controller processes before starting one.
The existing Node.js/Web filesync works the same way as before and not migrated to the new way in this PR.