cli: do nsboot bootstrap very early in the ns execution.
The idea here is that we want to minimize the amount of cycles spent in the old
`ns`, before executing the new one.

Also removed the fallback download when entering from `ns`, as if we're in `ns`
then we already have a binary.

Unfortunately also had to remove the `--autoupdate` flag, as this executes
before flag parsing. Disabling auto-updating will have to be done exclusively
via environment or the configuration.