ns: make sure that filewatcher.SetupFileWatcher() is invoked.
This addresses the regression reported in #1100. A recent change removed file
watching registration from the main cli entrypoint to allow us to introduce new
clis. In the process, this critical bit of initialization was left out.

Because this is exercised only during `ns dev`, it was not caught by our tests.