Expose all HTTP response headers from gRPC-Transcoding proxy to CORS callers.
This allows webui to observe grpc-message response header and expose readable error description to a human.

By default browsers only expose a small set of response headers to CORS callers. Access-Control-Expose-Headers reponse header controls which headers are exposed. Here we set it to "*" (all) to allow the JS
app to access any header.