provider/vault: add providers for PEM data.
These providers take an existing secret reference that generate the TLS bundle, extract the PEM data and store them in a new secret. The use case is applications that expect PEM certificates and keys to be stored directly, e.g. in a filesystem. These can now be configured using configurable mounts pointing to secret bindings.