internal/ssh: use SSH_AUTH_SOCK env var when it is set
Currently in order to use ssh-agent, necessary to set agent socket path
in a remote configuration.
But when you want to use different authentication methods locally and in
CI, then different configurations required, which is not convenient as
whole workspace configuration needs to be duplicated.
With this patch we are closer to a default behaviour of most of SSH
tools.